refactor(IntroSection): use capital `I` `<IntroSection>` `IntroSection.tsx`
