+++ components/IntroSection.tsx
