+++ components/medLrgDevices.tsx
@@ -1,7 +1,7 @@
 import ReactFullpage from "@fullpage/react-fullpage"
 import Image, { StaticImageData } from "next/image"
 import { useEffect, useState } from "react"
-import IntroSection from "@/components/introSection"
+import IntroSection from "@/components/IntroSection"
 import AboutSection from "@/components/aboutSection"
 import WorkExperienceSection from "@/components/workExperienceSection"
 import Portfolio from "@/components/Portfolio"
