+++ pages/index.tsx
@@ -25,7 +25,7 @@ import Image, { StaticImageData } from "next/image"
 import { useEffect, useState } from "react"
 import Layout from "@/components/layout"
 import Navbar from "@/components/navbar"
-import IntroSection from "@/components/introSection"
+import IntroSection from "@/components/IntroSection"
 import AboutSection from "@/components/aboutSection"
 import TechStackSection from "@/components/techStackSection"
 import WorkExperienceSection from "@/components/workExperienceSection"
